refactor: extract wallet provider logic into focused hooks (#61)
Break up the monolithic WalletProvider by extracting five inlined
responsibilities into four dedicated hooks:

- useWalletPersistence — localStorage save/restore, Supabase profile sync
- useWalletBalance — Pollar/Horizon balance fetch, refresh, tx history
- usePollarWallet — Pollar login modal, Supabase auth sync
- useExternalWallet — Stellar Wallets Kit connect flow

WalletProvider is now a thin composition shell (~130 lines) that wires
the hooks together. Public API (WalletProvider, useWalletContext) is
unchanged.

Closes #61
